Task 6

+++ 0x08-recursion/6-is_prime_number.c
@@ -7,7 +7,7 @@
  */
 int is_prime_number(int n)
 {
-	long long i = 2;
+	unsigned long i = 2;
 
 	if (n == 0 || n <= 1)
 		return (0);
